# Improve performance by swapping nested loops