# Description
We were doing:
```
for each item in slice:
    for key, value in map:
```

This PR switches it to:
```
for key, value in map:
    for each item in slice:
```
Because iterating maps in Go has more overhead than iterating slices, so we want to reduce the amount of map iterations.

# Testing
Tested on 100 million records:
```
* | eval height=ResolutionHeight / 10 | stats avg(height) as avg, count(height) as count, sum(height) as sum by ResolutionWidth | sort -ResolutionWidth
```
This PR decreased the total real time from 10.8 seconds to 9.7 seconds. The CPU time of `readAllRawRecords()` dropped from 6.7 seconds to 3.3 seconds.